requirements.txt
Django==3.1
psycopg2>=2.8
djangorestframework
markdown
django-filter
coverage==3.6